Fixed thread safety issues on using CTPP2 template formatter (BRAINSTORM-1432)
PluginData and StoredQueryHandlerBase cached template formatters which are
not thread safe. The cached formatters were available to all query threads,
and hence the code was not thread safe, since the formatter itself is not
thread safe. The TemplateFactory provides thread safe access to the formatter
by using a thread local cache. Fixed PluginData and StoredQueryHandlerBase
not to cache the formatters, but a shared pointer to one is requested from
the template factory separately for each request.